debootstrap a default variant (instead of the minbase variant)
It was reported that cron is not installed in the Kali Cloud image.
Seems like some users expect it to be there, and it's true that cron is
a very standard tool that has been available pretty much forever. It's
also true that cron is installed in other Kali images, so why is it
missing from the Cloud image?

Looking at the details, cron is a package with a priority of
"Important". However debootstrap is used with `--variant=minbase`, so
important packages are treated as any other packages, and they are not
necessarily installed. It depends if they are explicitly listed for
installation (eg. in a Kali metapackage) or if other packages depend on
it.

However cron is not explicitly listed in any metapackage, and most
packages that depend or recommend cron do so with 'cron | systemd', as
systemd provides a similar functionality via systemd timers.

In the end, and since systemd is installed, apt doesn't see it as
necessary to install cron.

It seems that if we want cron to be installed, the easiest is just to
drop `--variant=minbase`.

After doing two consecutive builds, one with minbase, one without, the
difference in packages installed is as follow:

    ii  bsd-mailx
            ii  cron
            ii  debconf-i18n
    ii  exim4-base
    ii  exim4-config
    ii  exim4-daemon-light
    ii  libatm1:amd64
    ii  libgnutls-dane0:amd64
    ii  liblockfile1:amd64
    ii  liblockfile-bin
    ii  libnss-systemd:amd64
    ii  libpam-cap:amd64
            ii  libtext-charwidth-perl
            ii  libtext-iconv-perl
            ii  libtext-wrapi18n-perl
    ii  libunbound8:amd64

Column 1 are packages unique to the build with `--variant=minbase`, column
2 are packages unique to the build without `minbase`.

We can see that removing `--variant=minbase` has the consequence of
installing `cron` and `debconf-i18n`, along with its dependencies
`libtext-*`. However, it also leads to the removal of `exim4` and its
dependencies.

As a result, the image produced WITHOUT `--variant=minbase` is a bit
smaller than the minbase variant (the tar file is 5MB less).  It's a bit
counter-intuitive, I know.