Base Kali config on Bullseye rather than Sid
There are important changes in the network config in Debian Sid
(systemd-networkd, systemd-resolved, netplan), and at a quick glance, it
broke the network for Kali on EC2.

Looking at the git history, and for example commit such as
de6453f68f243bc30ee172a78819a9929387db06, I'm not even sure that the new
network setup has been tested and is functional for EC2. This commit
added netplan config for the GCE and GENERIC image, does that mean that
there's no need for config on EC2, or does that mean that EC2 was not
taken care of yet? I do not know.

So let's hold back on those changes for now, and stick to the Bullseye
network config.