Fix AZURE/999-python2
Let the first command fail. Honestly I don't know why it fails, the
output (stdout or sterr) doesn't appear in the logs.

I guess that the test on usr/bin/python was meant to ensure that the
command was expected to succeed, but since it doesn't ensure that,
let's remove it then.